Replace `std::fs` calls with `fs_err`
By default `std::fs` does not include debugging information such as the filename. The `fs_err` crate overcomes this by wrapping all `std::fs` calls and returning a normal `std::io::Error` but with the additional information.

We should always provide this information to customers and other buildpack developers when it's safe to do so (which for a buildpack author is basically always).