Use bullet_stream WIP test feature
This feature globally blocks all other threads from modifying the writer for the closure's duration. The prior thread_local based logger has a race condition when "paragraph" text is being used and many tests are calling `global::set_version` as every time that is called, the state of the text that was just written is lost (this is used to track whether or not we need a newline).

This commit validates usage of capturing build output in tests. Once all maintainers are satisfied with the feature, it will be released and we can update this repo to use the released version.