Use function print interface (#115)
* Update to latest detect style (and global)

Guidance is here:

https://github.com/heroku-buildpacks/bullet_stream/blob/fbfa7b377542eb6a550861a525cac03ca61c7180/examples/style_guide.rs#L21-L36

* Use global "all done" message

* Global header print

* Update method return type

* Remove logging variable

* Replace streaming interface with functions

* Replace streaming interface with functions

* Update bullet_stream

* Use function style output API

* Use function style output API

* Remove unused argument from function

* Use function interface

* Update log interface, change function args

* Change return signature

* Prefer function interface

* Prefer function interface

* Prefer function interface

* Prefer function interface

* Prefer function interface

* Prefer function interface

* Prefer function interface

* Prefer function interface

* Prefer function interface

* Update imports

* Integration tests stdout -> stderr

* Fix integration test

The python buildpack is still printing to stdout, so this test fails to find the expected line. Combining stdout and stderr for the assertion fixes the tests.

* Prefer function interface

* Update function args

* Fix error logging tests

* Remove unused crate

* Use bullet_stream WIP test feature

This feature globally blocks all other threads from modifying the writer for the closure's duration. The prior thread_local based logger has a race condition when "paragraph" text is being used and many tests are calling `global::set_version` as every time that is called, the state of the text that was just written is lost (this is used to track whether or not we need a newline).

This commit validates usage of capturing build output in tests. Once all maintainers are satisfied with the feature, it will be released and we can update this repo to use the released version.

* Use stdout+stderr in tests

I'm honestly not sure why this is failing. I think that `pack` is doing some output redirection.

* Update BS branch

* Use released bullet_stream crate

* Fix variable name